restorer: move zeroPrefixLen to restic package
