+++ internal/restorer/sparsewrite.go
@@ -3,7 +3,9 @@
 
 package restorer
 
-import "bytes"
+import (
+	"github.com/restic/restic/internal/restic"
+)
 
 // WriteAt writes p to f.File at offset. It tries to do a sparse write
 // and updates f.size.
@@ -16,7 +18,7 @@ func (f *partialFile) WriteAt(p []byte, offset int64) (n int, err error) {
 
 	// Skip the longest all-zero prefix of p.
 	// If it's long enough, we can punch a hole in the file.
-	skipped := zeroPrefixLen(p)
+	skipped := restic.ZeroPrefixLen(p)
 	p = p[skipped:]
 	offset += int64(skipped)
 
@@ -32,22 +34,4 @@ func (f *partialFile) WriteAt(p []byte, offset int64) (n int, err error) {
 	}
 
 	return n, err
-}
-
-// zeroPrefixLen returns the length of the longest all-zero prefix of p.
-func zeroPrefixLen(p []byte) (n int) {
-	// First skip 1kB-sized blocks, for speed.
-	var zeros [1024]byte
-
-	for len(p) >= len(zeros) && bytes.Equal(p[:len(zeros)], zeros[:]) {
-		p = p[len(zeros):]
-		n += len(zeros)
-	}
-
-	for len(p) > 0 && p[0] == 0 {
-		p = p[1:]
-		n++
-	}
-
-	return n
 }
